📁 Finalizing makefile
